requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.1.0
Babel==2.15.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
branca==0.7.2
build==1.2.2.post1
cachetools==5.5.2
category-encoders==2.6.4
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
chromadb==1.0.10
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
docopt==0.6.2
durationpy==0.10
esda==2.6.0
et_xmlfile==2.0.0
executing==2.0.1
faiss-cpu==1.11.0
fastapi==0.115.9
fastjsonschema==2.20.0
filelock==3.16.0
findspark==2.0.1
Flask==3.1.1
flatbuffers==24.3.25
folium==0.17.0
fonttools==4.53.1
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.7.0
fsspec==2024.9.0
gast==0.6.0
geopandas==1.0.1
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.1
huggingface-hub==0.31.4
humanfriendly==10.0
idna==3.7
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.6.0
jaxlib==0.6.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.10.0
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kaleido==0.2.1
keras==3.5.0
kiwisolver==1.4.5
kubernetes==32.0.1
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.67
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
langsmith==0.4.4
libclang==18.1.1
libpysal==4.12.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.0
matplotlib-inline==0.1.7
matplotlib-venn==1.1.1
mdurl==0.1.2
mgwr==2.2.1
mistune==3.0.2
ml_dtypes==0.5.1
mmh3==5.1.0
mpmath==1.3.0
multidict==6.6.3
multitasking==0.0.11
mypy_extensions==1.1.0
mysql-connector-python==9.0.0
namex==0.0.8
narwhals==1.26.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.82.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.18
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pdf2image==1.17.0
peewee==3.17.6
pillow==10.4.0
pipenv==2025.0.3
platformdirs==4.2.2
plotly==6.0.0
posthog==4.0.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.0.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.10.1
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.26.3
pyogrio==0.9.0
pyparsing==3.1.2
pypdf==5.7.0
PyPika==0.48.9
pyproj==3.6.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyspark==3.5.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.1.0
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.11.6
reportlab==4.4.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.19.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==72.1.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.5
spglm==1.1.0
spreg==1.5.0
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.45.3
statsmodels==0.14.2
stemgraphic==0.9.1
streamlit==1.43.1
streamlit_folium==0.24.0
sympy==1.13.2
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-intel==2.17.0
termcolor==2.4.0
terminado==0.18.1
tf_keras==2.19.0
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.21.1
toml==0.10.2
torch==2.4.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.0
typer==0.15.4
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.34.2
virtualenv==20.31.2
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.44.0
widgetsnbextension==4.0.11
wrapt==1.16.0
XlsxWriter==3.2.2
xyzservices==2024.6.0
yarl==1.20.1
yfinance==0.2.43
zipp==3.21.0
zstandard==0.23.0
